+++ extensions-core/multi-stage-query/src/main/java/org/apache/druid/msq/input/external/ExternalInputSpecSlicer.java
@@ -19,8 +19,8 @@
 
 package org.apache.druid.msq.input.external;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Iterators;
+import org.apache.druid.data.input.FilePerSplitHintSpec;
 import org.apache.druid.data.input.InputFileAttribute;
 import org.apache.druid.data.input.InputSource;
 import org.apache.druid.data.input.InputSplit;
@@ -161,7 +161,7 @@ public class ExternalInputSpecSlicer implements InputSpecSlicer
         ((SplittableInputSource<?>) splittableInputSource.withSplit((InputSplit) split))
             .createSplits(spec.getInputFormat(), FilePerSplitHintSpec.INSTANCE)
             .map(subSplit -> splittableInputSource.withSplit((InputSplit) subSplit))
-            .forEach(s -> ((List) subSources).add(s));
+            .forEach(subSources::add);
       }
 
       if (subSources.isEmpty()) {
@@ -236,29 +236,6 @@ public class ExternalInputSpecSlicer implements InputSpecSlicer
           ).iterator(),
           xs -> !xs.isEmpty()
       );
-    }
-  }
-
-  /**
-   * Assigns each input file to its own split.
-   */
-  @VisibleForTesting
-  static class FilePerSplitHintSpec implements SplitHintSpec
-  {
-    static FilePerSplitHintSpec INSTANCE = new FilePerSplitHintSpec();
-
-    private FilePerSplitHintSpec()
-    {
-      // Singleton.
-    }
-
-    @Override
-    public <T> Iterator<List<T>> split(
-        final Iterator<T> inputIterator,
-        final Function<T, InputFileAttribute> inputAttributeExtractor
-    )
-    {
-      return Iterators.transform(inputIterator, Collections::singletonList);
     }
   }
 }
